requirements.txt
numpy>=1.18.0
gitpython>=3.1.0
rich>=10.0.0
click>=7.0.0
python-rapidjson>=1.5
